Changes in the destination folder and playlist
Changed the destination to a folder from Desktop to a folder of your
choosing by displaying a broswe dialog.

Changed the script so that can now exports the selected playlist,
instead of the current playlist. The difference is this, you're
listening to a song but you want to export another playlist, the
selected playist. With the method "current playlist" the script will
export the playlist of the track that currently playing. Instead with
the changes made in this commit the script will export the actual
selected playlist.